experiment params updated. installed typesafe config. forces rewrite of command line parsing.

+++ src/main/scala/cse/fitzgero/sorouting/app/MATSimSnapshotGenerator.scala
@@ -6,7 +6,7 @@ object MATSimSnapshotGenerator extends App {
 
   // example AppConfig("examples/tutorial/programming/example7-config.xml", "output/example7", "5", "06:00:00", "07:00:00", ArgsNotMissingValues)
   val appConfig: MATSimRunnerConfig = args match {
-    case Array(in, out, wD, sT, eT) => MATSimRunnerConfig(in, out, wD, sT, eT, ArgsNotMissingValues)
+    case Array(in, out, wD, sT, eT) => MATSimRunnerConfig(in, out, wD.toInt, sT, eT, ArgsNotMissingValues)
     case _ => MATSimRunnerConfig()
   }
